fix: create lib folder which only implements the facade pattern of libraries
